+++ PLViOSMediaPlayerDemo/PolyvIOSMediaPlayerDemo/PolyvVodScenes/Common/GeneralUI/PLVSubtitle/UI/PLVVodMediaSubtitleManager.h
@@ -1,20 +1,19 @@
 //
-//  PLVVodSubtitleManager.h
-//  PLVVodSubtitleDemo
+//  PLVVodMediaSubtitleManager.h
+//  PolyvIOSMediaPlayerDemo
 //
-//  Created by Bq Lin on 2017/12/4.
-//  Copyright © 2017年 POLYV. All rights reserved.
+//  Created by polyv on 2024/12/23.
 //
 
-#import <UIKit/UIKit.h>
-#import "PLVVodSubtitleItem.h"
-#import "PLVVodSubtitleViewModel.h"
+#import <Foundation/Foundation.h>
+#import "PLVVodMediaSubtitleItem.h"
+#import "PLVVodMediaSubtitleViewModel.h"
 
-@interface PLVVodSubtitleManager : NSObject
+@interface PLVVodMediaSubtitleManager : NSObject
 
-@property (nonatomic, strong, readonly) NSMutableArray<PLVVodSubtitleItem *> *subtitleItems;
+@property (nonatomic, strong, readonly) NSMutableArray<PLVVodMediaSubtitleItem *> *subtitleItems;
 
-@property (nonatomic, strong, readonly) NSMutableArray<PLVVodSubtitleItem *> *subtitleItems2;
+@property (nonatomic, strong, readonly) NSMutableArray<PLVVodMediaSubtitleItem *> *subtitleItems2;
 
 // 仅底部字幕 单字幕模式
 + (instancetype)managerWithSubtitle:(NSString *)subtitle label:(UILabel *)subtitleLabel error:(NSError **)error;
@@ -25,7 +24,7 @@
 - (void)showSubtitleWithTime:(NSTimeInterval)time;
 
 // 底部字幕+顶部字幕 单字幕模式支持样式自定义
-+ (instancetype)managerWithSubtitle:(NSString *)subtitle style:(PLVVodSubtitleItemStyle *)style label:(UILabel *)subtitleLabel topLabel:(UILabel *)subtitleTopLabel error:(NSError **)error;
++ (instancetype)managerWithSubtitle:(NSString *)subtitle style:(PLVVodMediaSubtitleItemStyle *)style label:(UILabel *)subtitleLabel topLabel:(UILabel *)subtitleTopLabel error:(NSError **)error;
 
 /// 底部+顶部字幕 支持单字幕/双字幕模式
 ///
@@ -39,6 +38,6 @@
 /// @param style2 字幕样式2
 /// @param subtitleLabel2 底部字幕（上）
 /// @param subtitleTopLabel2 顶部字幕（下）
-+ (instancetype)managerWithSubtitle:(NSString *)subtitle style:(PLVVodSubtitleItemStyle *)style error:(NSError **)error subtitle2:(NSString *)subtitle2 style2:(PLVVodSubtitleItemStyle *)style2  error2:(NSError **)error2 label:(UILabel *)subtitleLabel topLabel:(UILabel *)subtitleTopLabel label2:(UILabel *)subtitleLabel2 topLabel2:(UILabel *)subtitleTopLabel2;
++ (instancetype)managerWithSubtitle:(NSString *)subtitle style:(PLVVodMediaSubtitleItemStyle *)style error:(NSError **)error subtitle2:(NSString *)subtitle2 style2:(PLVVodMediaSubtitleItemStyle *)style2  error2:(NSError **)error2 label:(UILabel *)subtitleLabel topLabel:(UILabel *)subtitleTopLabel label2:(UILabel *)subtitleLabel2 topLabel2:(UILabel *)subtitleTopLabel2;
 
 @end